MR-65: Re-formated code to use routes in Layout
