+++ src/Components/Shared/ConfirmModal/modalConfirm.module.css
@@ -1,6 +1,6 @@
 .modal {
   position: fixed;
-  z-index: 1;
+  z-index: 999;
   left: 0;
   top: 0;
   width: 100%;
