+++ src/Components/SuperAdmins/Form/form.module.css
@@ -1,16 +1,5 @@
-.formBackground{
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: #00000080;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 9999;
-}
 .container{
+  margin: auto;
   width: 35vw;
   background-color: white;
   box-shadow: 0px 10px 23px #1E1E1E;
