+++ src/Components/SuperAdmins/index.jsx
@@ -1,23 +1,16 @@
 import styles from './super-admins.module.css';
-import Form from './Form/index';
 import ResponseModal from '../Shared/ResponseModal';
 import Table from './Table/index';
 import { useEffect, useState } from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import ConfirmModal from '../Shared/ConfirmModal';
 
 function SuperAdmins() {
   const [superadmins, setSuperadmins] = useState([]);
   const [showModal, setshowModal] = useState(false);
-  const [showConfirmAdd, setShowConfirmAdd] = useState(false);
-  const [showConfirmEdit, setShowConfirmEdit] = useState(false);
   const [showConfirmDelete, setShowConfirmDelete] = useState(false);
   const [idToDelete, setIdToDelete] = useState(null);
-  const [superadminToAdd, setSuperadminToAdd] = useState(null);
-  const [idToEdit, setIdToEdit] = useState(null);
-  const [editedSuperadmin, setEditedSuperadmin] = useState(null);
   const [resMessage, setResMessage] = useState('');
-  const [state, setSate] = useState('');
+  const [state, setState] = useState('');
 
   const getSuperadmins = async () => {
     try {
@@ -25,81 +18,28 @@ function SuperAdmins() {
       const data = await res.json();
       setSuperadmins(data.data);
     } catch (error) {
-      setSate('fail');
+      setState('fail');
       setResMessage(error.message);
       openModal();
     }
   };
-  useEffect(() => {
-    getSuperadmins();
-  });
-
   const closeModal = () => {
     setshowModal(false);
+    sessionStorage.clear();
   };
   const openModal = () => {
     setshowModal(true);
   };
-  const confirmAdd = async (newSuperadmin) => {
-    setShowConfirmAdd(true);
-    setSuperadminToAdd(newSuperadmin);
-  };
-  const addItem = async () => {
-    try {
-      const res = await fetch(`${process.env.REACT_APP_API_URL}/api/super-admins`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(superadminToAdd)
-      });
-      setShowConfirmAdd(false);
-      getSuperadmins();
-      if (res.ok) {
-        setSate('success');
-        setResMessage('New superadmin created');
-        openModal();
-      } else {
-        setSate('fail');
-        setResMessage('Failed to create superadmin');
-        openModal();
-      }
-    } catch (error) {
-      setSate('fail');
-      setResMessage(error.message);
-      openModal();
-    }
-  };
-  const confirmEdit = (id, updatedSuperadmin) => {
-    setShowConfirmEdit(true);
-    setIdToEdit(id);
-    setEditedSuperadmin(updatedSuperadmin);
-  };
-  const putItem = async () => {
-    try {
-      const res = await fetch(`${process.env.REACT_APP_API_URL}/api/super-admins/${idToEdit}`, {
-        method: 'PUT',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(editedSuperadmin)
-      });
-      setShowConfirmEdit(false);
-      if (res.ok) {
-        setSate('success');
-        setResMessage('Superadmin updated');
-        openModal();
-      } else {
-        setSate('fail');
-        setResMessage('Failed to update superadmin');
-        openModal();
-      }
-    } catch (error) {
-      setSate('fail');
-      setResMessage(error.message);
-      openModal();
-    }
-  };
+  useEffect(() => {
+    const getResponseFromForm = () => {
+      setState(sessionStorage.getItem('state'));
+      setResMessage(sessionStorage.getItem('resMessage'));
+      sessionStorage.getItem('state') && openModal();
+    };
+    getResponseFromForm();
+    getSuperadmins();
+  });
+
   const confirmDelete = (id) => {
     setShowConfirmDelete(true);
     setIdToDelete(id);
@@ -111,46 +51,22 @@ function SuperAdmins() {
       });
       setShowConfirmDelete(false);
       if (res.ok) {
-        setSate('success');
-        setResMessage('Superadmin deleted');
-        openModal();
+        sessionStorage.setItem('state', 'success');
+        sessionStorage.setItem('resMessage', 'Superadmin deleted');
         setSuperadmins([...superadmins.filter((superadmin) => superadmin._id !== idToDelete)]);
       } else {
-        setSate('fail');
-        setResMessage('Failed to delete superadmin');
-        openModal();
+        sessionStorage.setItem('state', 'fail');
+        sessionStorage.setItem('resMessage', 'Failed to delete superadmin');
       }
     } catch (error) {
-      setSate('fail');
-      setResMessage(error.message);
-      openModal();
+      sessionStorage.setItem('state', 'fail');
+      sessionStorage.setItem('resMessage', 'Failed to delete superadmin');
     }
   };
 
   return (
     <section className={styles.container}>
       {showModal && <ResponseModal state={state} message={resMessage} handler={closeModal} />}
-
-      {showConfirmAdd && (
-        <ConfirmModal
-          title={'New superadmin'}
-          reason={'submit'}
-          handler={() => setShowConfirmAdd(false)}
-          onAction={addItem}
-        >
-          Are you sure you want to add this superadmin?
-        </ConfirmModal>
-      )}
-      {showConfirmEdit && (
-        <ConfirmModal
-          title={'Edit superadmin'}
-          reason={'submit'}
-          handler={() => setShowConfirmEdit(false)}
-          onAction={putItem}
-        >
-          Are you sure you want to edit this superadmin?
-        </ConfirmModal>
-      )}
       {showConfirmDelete && (
         <ConfirmModal
           title={'Delete superadmin'}
@@ -162,17 +78,7 @@ function SuperAdmins() {
         </ConfirmModal>
       )}
       <h2 className={styles.h2}>Superadmin List</h2>
-      <Router>
-        <Table data={superadmins} deleteItem={deleteItem} confirmDelete={confirmDelete} />
-        <Switch>
-          <Route path="/super-admins/form/:id">
-            <Form putItem={putItem} addItem={addItem} confirmEdit={confirmEdit} />
-          </Route>
-          <Route path="/super-admins/form">
-            <Form addItem={addItem} confirmAdd={confirmAdd} />
-          </Route>
-        </Switch>
-      </Router>
+      <Table data={superadmins} confirmDelete={confirmDelete} />
     </section>
   );
 }
